+++ vulnshield-iso/VulnScan.Web/Program.cs
@@ -89,6 +89,7 @@ builder.Services.AddScoped<IAutoImportService, AutoImportService>();
 builder.Services.AddScoped<IGreenboneGmpClient, GreenboneGmpClient>();
 builder.Services.AddScoped<IGreenboneImportService, GreenboneImportService>();
 builder.Services.AddScoped<IGreenboneSettingsService, GreenboneSettingsService>();
+builder.Services.AddScoped<ISystemCheckService, SystemCheckService>();
 builder.Services.AddScoped<IScanAllowedRangeService, ScanAllowedRangeService>();
 builder.Services.AddScoped<INmapService, NmapService>();
 builder.Services.AddScoped<INmapXmlParserService, NmapXmlParserService>();